+++ dask_ml/cluster/spectral.py
@@ -190,7 +190,8 @@ class SpectralClustering(BaseEstimator, ClusterMixin):
         if isinstance(self.assign_labels, six.string_types):
             if self.assign_labels == "kmeans":
                 km = KMeans(
-                    n_clusters=n_clusters, random_state=draw_seed(rng, 2 ** 32 - 1)
+                    n_clusters=n_clusters,
+                    random_state=draw_seed(rng, 2 ** 32 - 1, dtype="uint"),
                 )
             elif self.assign_labels == "sklearn-kmeans":
                 km = sklearn.cluster.KMeans(n_clusters=n_clusters, random_state=rng)
